+++ components/theme/index.ts
@@ -8,7 +8,7 @@ import {
   getOwner,
 } from 'solid-js';
 import { colorParse, mixColor, toneColor } from '@moneko/common';
-import { css } from '@moneko/css';
+import { CSSObject, css } from '@moneko/css';
 export { toneColor } from '@moneko/common';
 
 /** 生成主题色调
@@ -16,7 +16,7 @@ export { toneColor } from '@moneko/common';
  * @param {ThemeOption} option 配置项
  * @returns {Record<string, string>} 主题色调
  */
-export function generateTheme(base: string, option: ThemeOption): Record<string, string> {
+export function generateColor(base: string, option: ThemeOption): Record<string, string> {
   const obj = toneColor(base, option?.dark);
   const textColor = colorParse(obj[5]);
   const baseColor = colorParse(base);
@@ -27,25 +27,20 @@ export function generateTheme(base: string, option: ThemeOption): Record<string,
     [`--${option.name}-heading`]: obj[5],
     [`--${option.name}-active`]: obj[30],
     [`--${option.name}-color`]: obj[40],
-    // [`--${option.name}-hover`]: obj[50],
     [`--${option.name}-hover`]: mixColor(obj[40], obj[30], 15),
     [`--${option.name}-secondary-bg`]: obj[70],
     [`--${option.name}-border`]: obj[80],
-    // [`--${option.name}-outline`]: obj[90],
     [`--${option.name}-outline`]: mixColor(obj[90], obj[40], 5),
     [`--${option.name}-selection`]: obj[90],
     [`--${option.name}-notify-bg`]: colorParse(obj[90]).setAlpha(0.8).toRgbaString(),
     [`--on-${option.name}-selection`]: obj[10],
     [`--${option.name}-shadow`]: baseColor.setAlpha(0.12).toRgbaString(),
+    [`--${option.name}-shadow-1`]: baseColor.setAlpha(0.08).toRgbaString(),
+    [`--${option.name}-shadow-2`]: baseColor.setAlpha(0.05).toRgbaString(),
     [`--${option.name}-details-bg`]: obj[95],
     [`--${option.name}-component-bg`]: obj[98],
     [`--${option.name}-bg`]: obj[99],
     [`--on-${option.name}-bg`]: obj[5],
-    [`--${option.name}-base-shadow`]: `0 3px 6px -4px ${baseColor
-      .setAlpha(0.12)
-      .toRgbaString()}, 0 6px 16px 0 ${baseColor
-      .setAlpha(0.08)
-      .toRgbaString()}, 0 9px 28px 8px ${baseColor.setAlpha(0.05).toRgbaString()}`,
   };
 }
 
@@ -60,6 +55,43 @@ export enum ColorScheme {
 }
 
 function createTheme() {
+  const baseTokens = css({
+    ':root,:host': {
+      '--scrollbar-size': '4px',
+      '--font-size': '14px',
+      '--font-size-sm': '12px',
+      '--font-size-xs': '10px',
+      '--font-size-lg': '16px',
+      '--border-base': '1px solid var(--border-color)',
+      '--border-radius': '8px',
+      '--text-color': 'var(--primary-text)',
+      '--text-secondary': 'var(--primary-secondary)',
+      '--text-heading': 'var(--primary-heading)',
+      '--text-selection': 'var(--primary-selection)',
+      '--transition-duration': '0.3s',
+      '--mask-bg': 'rgb(0 0 0 / 5%)',
+      '--modal-component-bg': 'rgb(255 255 255 / 80%)',
+      '--modal-box-shadow': '0 5px 35px rgb(0 0 0 / 10%)',
+      '--notification-box-shadow': '0 4px 16px rgb(0 0 0 / 5%)',
+      '--transition-timing-function': 'cubic-bezier(0.645, 0.045, 0.355, 1)',
+      fontSize: 'var(--font-size)',
+      fontFamily:
+        "-apple-system, BlinkMacSystemFont, 'Segoe UI', Roboto, 'Helvetica Neue', Arial, 'Noto Sans', sans-serif, 'Apple Color Emoji', 'Segoe UI Emoji', 'Segoe UI Symbol', 'Noto Color Emoji', Helvetica, Oxygen, Ubuntu, Cantarell, 'Fira Sans', 'Droid Sans'",
+      color: 'var(--text-color, rgb(0 0 0 / 65%))',
+      lineHeight: 1.8,
+    },
+    '[disabled]:not([disabled="false"])': {
+      cursor: 'not-allowed',
+      color: 'var(--disable-color)',
+    },
+    '::selection': {
+      backgroundColor: 'var(--text-selection)',
+    },
+    '::-webkit-scrollbar': {
+      inlineSize: 'var(--scrollbar-size)',
+      blockSize: 'var(--scrollbar-size)',
+    },
+  });
   const preset =
     ColorScheme[localStorage.getItem('color-scheme') as keyof typeof ColorScheme] || 'auto';
   /** 检测 prefers-color-scheme 媒体查询是否为 light 模式 */
@@ -78,102 +110,60 @@ function createTheme() {
     error: '#901c22',
     success: '#419418',
   });
-  const primary = createMemo(() => generateTheme(light().primary, { name: 'primary' }));
-  const warning = createMemo(() => generateTheme(light().warning, { name: 'warning' }));
-  const success = createMemo(() => generateTheme(light().success, { name: 'success' }));
-  const error = createMemo(() => generateTheme(light().error, { name: 'error' }));
-  const lightCss = createMemo(() => {
-    return css`
-      :root,
-      :host {
-        ${primary()}
-        ${warning()}
-        ${success()}
-        ${error()}
-        --bg: transparent;
-        --disable-color: rgb(0 0 0 / 25%);
-        --disable-bg: rgb(0 0 0 / 4%);
-        --disable-border: #d9d9d9;
-        --border-color: var(--primary-border);
-        --component-bg: var(--primary-bg);
-      }
-    `;
-  });
+  const primary = createMemo(() => generateColor(light().primary, { name: 'primary' }));
   const darkPrimary = createMemo(() =>
-    generateTheme(dark().primary, { name: 'primary', dark: true }),
+    generateColor(dark().primary, { name: 'primary', dark: true }),
   );
+  const warning = createMemo(() => generateColor(light().warning, { name: 'warning' }));
   const darkWarning = createMemo(() =>
-    generateTheme(dark().warning, { name: 'warning', dark: true }),
+    generateColor(dark().warning, { name: 'warning', dark: true }),
   );
+  const success = createMemo(() => generateColor(light().success, { name: 'success' }));
   const darkSuccess = createMemo(() =>
-    generateTheme(dark().success, { name: 'success', dark: true }),
+    generateColor(dark().success, { name: 'success', dark: true }),
   );
-  const darkError = createMemo(() => generateTheme(dark().error, { name: 'error', dark: true }));
-  const darkCss = createMemo(() => {
-    return css`
-      :root,
-      :host {
-        ${darkPrimary()}
-        ${darkWarning()}
-        ${darkError()}
-        ${darkSuccess()}
-        --bg: #1c1c1c;
-        --disable-color: rgb(255 255 255 / 25%);
-        --disable-bg: rgb(255 255 255 / 8%);
-        --disable-border: #424242;
-        --border-color: #303030;
-        --component-bg: #141414;
-        --primary-shadow: rgb(0 0 0 / 12%);
-        --primary-selection: rgb(255 255 255 / 5%);
-        --primary-details-bg: rgb(255 255 255 / 3%);
-        --primary-component-bg: #000;
-        --modal-component-bg: rgb(30 30 30 / 80%);
+  const error = createMemo(() => generateColor(light().error, { name: 'error' }));
+  const darkError = createMemo(() => generateColor(dark().error, { name: 'error', dark: true }));
+
+  function getHostCss(tokens: CSSObject) {
+    let str = '';
+
+    for (const key in tokens) {
+      if (Object.prototype.hasOwnProperty.call(tokens, key)) {
+        str += `${key}:${tokens[key]};`;
       }
-    `;
+    }
+    return `${baseTokens}:root,:host{${str}}`;
+  }
+  const lightCss = createMemo(() => {
+    const tokens = Object.assign({}, primary(), warning(), success(), error(), {
+      '--bg': 'transparent',
+      '--disable-color': 'rgb(0 0 0 / 25%)',
+      '--disable-bg': 'rgb(0 0 0 / 4%)',
+      '--disable-border': '#d9d9d9',
+      '--border-color': 'var(--primary-border)',
+      '--component-bg': 'var(--primary-bg)',
+    });
+
+    return getHostCss(tokens);
   });
-  const baseCss = css`
-    :root,
-    :host {
-      --font-size: 14px;
-      --font-size-sm: 12px;
-      --font-size-xs: 10px;
-      --font-size-lg: 16px;
-      --border-base: 1px solid var(--border-color);
-      --border-radius: 8px;
-      --text-color: var(--primary-text);
-      --text-secondary: var(--primary-secondary);
-      --text-heading: var(--primary-heading);
-      --text-selection: var(--primary-selection);
-      --box-shadow-base: var(--primary-base-shadow);
-      --transition-duration: 0.3s;
-      --mask-bg: rgb(0 0 0 / 5%);
-      --modal-component-bg: rgb(255 255 255 / 80%);
-      --modal-box-shadow: 0 5px 35px rgb(0 0 0 / 10%);
-      --notification-box-shadow: 0 4px 16px rgb(0 0 0 / 5%);
-      --transition-timing-function: cubic-bezier(0.645, 0.045, 0.355, 1);
+  const darkCss = createMemo(() => {
+    const tokens = Object.assign({}, darkPrimary(), darkWarning(), darkSuccess(), darkError(), {
+      '--bg': '#1c1c1c',
+      '--disable-color': 'rgb(255 255 255 / 25%)',
+      '--disable-bg': 'rgb(255 255 255 / 8%)',
+      '--disable-border': '#424242',
+      '--border-color': '#303030',
+      '--component-bg': '#141414',
+      '--primary-shadow': 'rgb(0 0 0 / 12%)',
+      '--primary-selection': 'rgb(255 255 255 / 5%)',
+      '--primary-details-bg': 'rgb(255 255 255 / 3%)',
+      '--primary-component-bg': '#000',
+      '--modal-component-bg': 'rgb(30 30 30 / 80%)',
+    });
 
-      font-size: var(--font-size);
-      font-family: -apple-system, BlinkMacSystemFont, 'Segoe UI', Roboto, 'Helvetica Neue', Arial,
-        'Noto Sans', sans-serif, 'Apple Color Emoji', 'Segoe UI Emoji', 'Segoe UI Symbol',
-        'Noto Color Emoji', Helvetica, Oxygen, Ubuntu, Cantarell, 'Fira Sans', 'Droid Sans';
-      color: var(--text-color, rgb(0 0 0 / 65%));
-      line-height: 1.8;
-    }
-
-    [disabled]:not([disabled='false']) {
-      cursor: not-allowed;
-      color: var(--disable-color);
-    }
-
-    ::selection {
-      background-color: var(--text-selection);
-    }
-
-    ::-webkit-scrollbar {
-      inline-size: 4px;
-      block-size: 4px;
-    }
-  `;
+    return getHostCss(tokens);
+  });
 
   function update(e: MediaQueryListEvent) {
     setIsDark(e.matches);
@@ -183,7 +173,7 @@ function createTheme() {
 
     setIsDark(_ === 'dark' || (_ === 'auto' && media.matches));
   });
-  const baseStyle = createMemo(() => baseCss + (isDark() ? darkCss() : lightCss()));
+  const baseStyle = createMemo(() => (isDark() ? darkCss() : lightCss()));
 
   createEffect(() => {
     if (scheme() === 'auto') {