+++ typings/lib.dom.d.ts
@@ -152,10 +152,5 @@ interface EyeDropper extends EyeDropperConstructor {
 
 interface Window {
   EyeDropper?: EyeDropper;
-  Prism: {
-    disableWorkerMessageHandler: boolean;
-    languages: Record<string, unknown>;
-    highlight(code: string, langs: unknown, lang: string): string;
-  };
   marked: typeof import('marked-completed');
 }